Move top right feed story menu into bottom action bar next to reblog and change icon to 3 dots icon, the same used in top right of single post
