+++ src/client/components/Story/Story.less
@@ -63,6 +63,7 @@
 
     &__text {
       margin: 0 6px;
+      width: 100%;
 
       @media @small {
         margin-left: 0 12px;
@@ -87,6 +88,10 @@
     top: -2px;
     font-size: 14px;
     color: #99aab5;
+  }
+
+  &__topics {
+    margin-left: auto;
   }
 
   &__topics > button {
