+++ src/client/components/StoryFooter/Buttons.less
@@ -67,4 +67,8 @@
       margin-left: 8px;
     }
   }
+
+  &__post-menu {
+    cursor: pointer;
+  }
 }
